refactor(write): Refactor in response to comments
